refactor(game-view): extract GameMapComponent
